chore(primitives): Move `ComponentClassName` to `@aws-amplify/ui` (#2751)
* Move `ComponentClassName` to `@aws-amplify/ui`

* Use ComponentClassNames instead in unit tests

* Create modern-spies-wave.md

* change filename
