+++ .changeset/modern-spies-wave.md
@@ -1,0 +1,6 @@
+---
+"@aws-amplify/ui-react": patch
+"@aws-amplify/ui": patch
+---
+
+(Internal patch): Move `ComponentClassName` to `@aws-amplify/ui`
