+++ packages/ui/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './authenticator';
+export * from './primitives';
 export * from './util';
