+++ packages/ui/src/types/primitives/index.ts
@@ -1,0 +1,1 @@
+export { ComponentClassName } from './componentClassName';
